cloud_function/requirements.txt
google-cloud-storage
paramiko
